resume-builder/linkedin_parser.py
import requests
from bs4 import BeautifulSoup
import re
import json
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from webdriver_manager.chrome import ChromeDriverManager
import time

class LinkedInParser:
    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        })
    
    def parse_profile(self, linkedin_url):
        """
        Parse LinkedIn profile data from URL
        For demo purposes, this will return mock data
        In production, you would need to implement proper LinkedIn scraping
        """
        try:
            # For demo purposes, return mock data
            # In a real implementation, you would need to:
            # 1. Handle LinkedIn authentication
            # 2. Use LinkedIn API or web scraping
            # 3. Respect LinkedIn's terms of service and rate limits
            
            return self._get_mock_profile_data()
            
        except Exception as e:
            print(f"Error parsing LinkedIn profile: {e}")
            return None
    
    def _get_mock_profile_data(self):
        """Return mock LinkedIn profile data for demonstration"""
        return {
            'personal_info': {
                'name': 'John Doe',
                'headline': 'Senior Software Engineer',
                'location': 'San Francisco, CA',
                'email': 'john.doe@email.com',
                'phone': '+1 (555) 123-4567',
                'linkedin_url': 'https://linkedin.com/in/johndoe',
                'summary': 'Experienced software engineer with 5+ years of expertise in full-stack development, specializing in Python, JavaScript, and cloud technologies. Passionate about building scalable applications and leading technical teams.'
            },
            'experience': [
                {
                    'title': 'Senior Software Engineer',
                    'company': 'Tech Corp',
                    'location': 'San Francisco, CA',
                    'duration': 'Jan 2022 - Present',
                    'description': 'Led development of microservices architecture, improved system performance by 40%, mentored junior developers, and implemented CI/CD pipelines.'
                },
                {
                    'title': 'Software Engineer',
                    'company': 'Startup Inc',
                    'location': 'San Francisco, CA',
                    'duration': 'Mar 2020 - Dec 2021',
                    'description': 'Developed full-stack web applications using React and Node.js, collaborated with cross-functional teams, and deployed applications to AWS.'
                },
                {
                    'title': 'Junior Developer',
                    'company': 'Digital Solutions',
                    'location': 'San Francisco, CA',
                    'duration': 'Jun 2019 - Feb 2020',
                    'description': 'Built responsive web applications, worked with REST APIs, and participated in code reviews and agile development processes.'
                }
            ],
            'education': [
                {
                    'degree': 'Bachelor of Science in Computer Science',
                    'school': 'University of California, Berkeley',
                    'location': 'Berkeley, CA',
                    'duration': '2015 - 2019',
                    'gpa': '3.8/4.0'
                }
            ],
            'skills': [
                'Python', 'JavaScript', 'React', 'Node.js', 'AWS', 'Docker',
                'Kubernetes', 'PostgreSQL', 'MongoDB', 'Git', 'REST APIs',
                'Microservices', 'CI/CD', 'Agile', 'Leadership'
            ],
            'certifications': [
                {
                    'name': 'AWS Certified Solutions Architect',
                    'issuer': 'Amazon Web Services',
                    'date': '2023'
                },
                {
                    'name': 'Certified Kubernetes Administrator',
                    'issuer': 'Cloud Native Computing Foundation',
                    'date': '2022'
                }
            ],
            'projects': [
                {
                    'name': 'E-commerce Platform',
                    'description': 'Built a scalable e-commerce platform using React, Node.js, and MongoDB. Implemented payment processing, inventory management, and analytics dashboard.',
                    'technologies': ['React', 'Node.js', 'MongoDB', 'Stripe API'],
                    'url': 'https://github.com/johndoe/ecommerce-platform'
                },
                {
                    'name': 'Task Management App',
                    'description': 'Developed a collaborative task management application with real-time updates, user authentication, and mobile-responsive design.',
                    'technologies': ['React', 'Firebase', 'Material-UI'],
                    'url': 'https://github.com/johndoe/task-manager'
                }
            ],
            'languages': [
                {'language': 'English', 'proficiency': 'Native'},
                {'language': 'Spanish', 'proficiency': 'Conversational'}
            ]
        }
    
    def _scrape_linkedin_profile(self, url):
        """
        Actual LinkedIn scraping implementation
        Note: This requires proper authentication and compliance with LinkedIn's terms
        """
        try:
            # Setup Chrome options for headless browsing
            chrome_options = Options()
            chrome_options.add_argument("--headless")
            chrome_options.add_argument("--no-sandbox")
            chrome_options.add_argument("--disable-dev-shm-usage")
            
            # Initialize driver
            driver = webdriver.Chrome(ChromeDriverManager().install(), options=chrome_options)
            
            # Navigate to LinkedIn profile
            driver.get(url)
            time.sleep(3)  # Wait for page to load
            
            # Extract profile data
            profile_data = {}
            
            # Extract name
            try:
                name_element = driver.find_element(By.CSS_SELECTOR, "h1.text-heading-xlarge")
                profile_data['name'] = name_element.text
            except:
                profile_data['name'] = "Name not found"
            
            # Extract headline
            try:
                headline_element = driver.find_element(By.CSS_SELECTOR, ".text-body-medium.break-words")
                profile_data['headline'] = headline_element.text
            except:
                profile_data['headline'] = "Headline not found"
            
            driver.quit()
            return profile_data
            
        except Exception as e:
            print(f"Error scraping LinkedIn: {e}")
            return None 