resume-builder/requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==0.8.11
Pillow>=11.3.0
jinja2==3.1.2
werkzeug==2.3.7
python-linkedin-v2==0.9.4
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0 